// src/data/node-types/car-models/types/CarModelNode.ts
import type {DataNodeType} from "../../../types/DataNodeType"

export type CarModelNode = {
    type: DataNodeType.CAR_MODEL
    data: {
        id: number
        name: string
        generation?: number
        internal_code?: string
        built_from?: number
        built_to?: number
        total_production?: number
        created_at: string
        updated_at: string
    }
}